requirements.txt
asgiref==3.11.0
dj-database-url==3.1.0
django==6.0.1
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
psycopg==3.3.2
psycopg-binary==3.3.2
pyjwt==2.11.0
python-dotenv==1.2.1
ruff==0.14.14
sqlparse==0.5.5
typing-extensions==4.15.0
whitenoise==6.11.0
